QScopedArrayPointer: document the most important ctor
The ctor that everyone would want to use was marked as \internal,
probably because of the SFINAE expression appearing in its signature.

Move the SFINAE to the template argument list, which QDoc hides from
the user, and drop the \internal. While at it, drop the home-grown
std::is_same re-implementation and use the real deal.

Change-Id: Ia357fe65f94e10ac9eeccb3490aa8b3e68114cbb
Reviewed-by: Lars Knoll <lars.knoll@qt.io>