xcb: remove DISPLAY_FROM_XCB macro
... as it does not add much value and is not widely used.

and cleanup code:

- where we repeatedly (for no good reason) call DISPLAY_FROM_XCB,
instead of storing display as a (member) variable.

- inconsistency where we (in the same function) alternatingly
use DISPLAY_FROM_XCB and variable holding pointer to Display
(when both refer to the same display).

In other places this patch replaces macros with code they would
translate to (or with minor adjustments to keep sensible line length).

Change-Id: Ieb2a3e055892dcf31f132891f83ed4a665e3fa6e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>