QKeySequence: let Private::decodeString() take QString by value
It calls toLower() on it, and the existence of two strings in scope
has already caused some latent bug where sizes of the two were mixed.

Taking by value means we can use the rvalue-version of toLower().
While that does not guarantee no allocations, we can make sure that at
least in QKeySequence::assign(), we hand in unshared data, by using
std::move().

Change-Id: Ic4a18486a6f4528eba36109dec1a47082cf57055
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>