qsimd_p.h: Don't set the __xxx__ variables with Clang and GCC
Commit 418184c2a0ad97cce12717a43f84fa6f12ece189 set some extra defines
that Clang and GCC do set so that MSVC and ICC builds would properly get
the features detected. But that meant we set them with Clang and GCC
(technically, set them again, but to the same value so no warning was
printed).

Don't do that. This commit allows me to use "-march=native -mno-rdrnd"
to disable the unconditional use of RDRAND instruction. That's required
to valgrind any applications, as the current version (3.12) does not
have support for that instruction.

 vex amd64->IR: unhandled instruction bytes: 0x48 0xF 0xC7 0xF0 0x48 0x8B 0x55 0xE8 0x48 0x89
 vex amd64->IR:   REX=1 REX.W=1 REX.R=0 REX.X=0 REX.B=0
 vex amd64->IR:   VEX=0 VEX.L=0 VEX.nVVVV=0x0 ESC=0F
 vex amd64->IR:   PFX.66=0 PFX.F2=0 PFX.F3=0
 ==78321== valgrind: Unrecognised instruction at address 0x4ef159c.
 ==78321==    at 0x4EF159C: _rdrand64_step (immintrin.h:208)
 ==78321==    by 0x4EF159C: qt_random_cpu(void*, long long) (qrandom.cpp:95)

Change-Id: Ia3e896da908f42939148fffd14c6884501de4fa4
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>