Blacklist AAAA related DNS lookup tests on Windows 7/8.1
These are failing in Qt 5.6 when using qemu/kvm and the tests
are skipped in 5.9 as Qt thinks IPv6 is not supported. The tests are
skipped when running under vmware (with vmxnet) for the same reason.
The actual reason for why the trip-A lookups are not working is unknown.
They do work when using nslookup in a command shell.

Task-number: QTQAINFRA-1201
Change-Id: I224f167be7d4657646ffcd16940cabec02f9d4e4
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>